% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamr.R
\name{compare_model}
\alias{compare_model}
\title{Compare summary statistics between various imputation methods}
\usage{
compare_model(df, feature, methods, missing_val_char)
}
\arguments{
\item{df}{A dataset with missing values that needs to be imputed.}

\item{feature}{(str) A string of column name, if the input data is a matrix, this should be a string like "Vn" where n is an integer representing the index of column}

\item{methods}{(str or list)-- the methods that users want to compare
Supporting methods are:
    CC     - Complete Case
    MIP - Imputation with mean value
    DIP - Imputation with median value}

\item{missing_val_char}{A string of a missing value format:
Supporting types are:
     NaN - Not a Number
     "" - Blank
     "?" - Question mark}
}
\value{
a summary table comparing the summary statistics: count, mean, std, min, 25\%, 50\%, 75\%, max.
}
\description{
This function will call function `impute_missing()` for several methods and
return a table with some statistical information of the specified feature
before and after imputation of different methods
}
\examples{
compare_model(data.frame(ex = c(1, 2, 3), bf = c(6, 8, "")), "bf", "DIP", "")

}
\seealso{
\code{\link{na.omit}} for the complete case

Other aggregate functions: \code{\link{impute_missing}}
}
\author{
Duong Vu, 2018
}
