% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamr-package.r
\docType{package}
\name{hamr}
\alias{hamr}
\alias{hamr-package}
\title{hamr: Plot, impute and compare missing values}
\description{
Provides a suite of functions to see missing value patterns, imputation with three simple methods and compare the results of the methods for a data frame or a matrix
}
