% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing.R
\name{impute_missing}
\alias{impute_missing}
\title{Impute missing values in a specified column of a data frame or a numerical matrix with three simple methods}
\usage{
impute_missing(dfm, col, method, missing_val_char)
}
\arguments{
\item{dfm}{A data frame or a numerical matrix}

\item{col}{A string of column name, if the input data is a matrix, this should be a string like "Vn" where n is an integer representing the index of column}

\item{method}{A string of a method name, should be one of "CC", "MIP" and "DIP"}

\item{missing_val_char}{A string of a missing value format, should be one of NA, NaN, "" and "?"}
}
\value{
A data frame having no missing values in the specified column
}
\description{
Impute missing values in a specified column of a data frame or a numerical matrix with three simple methods
}
\examples{
impute_missing(data.frame(ex = c(1, 2, 3), bf = c(6, 8, "")), "bf", "DIP", "")
impute_missing(matrix(c(1,2,3, 6,8,NA), nrow = 3, ncol = 2, byrow = FALSE), "V2", "DIP", NA)

}
\seealso{
\code{\link{na.omit}} for the complete case
}
