% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamr.R
\name{vis_missing}
\alias{vis_missing}
\title{Visualize missing data}
\usage{
vis_missing(df, colour = "default", missing_val_char = NA)
}
\arguments{
\item{df}{Input data frame or matrix}

\item{colour}{Colour scheme for plotting}

\item{missing_val_char}{the missing value character in the data frame, one of NA, "", " ", "?"}
}
\value{
Visualization of the missing data in a data set
}
\description{
Produces a visualization of all missing values in a data frame.
The missing values are encoded by the missing value character, which is NA by default.
}
\examples{
vis_missing(data.frame(ex = c(1, 2, 3), bf = c(6, 8, NA)))
}
\author{
Jordan Dubchak, March 2018
}
